test: Move `resolveDirZipBuffer` to global utils
