+++ node/packages/aws-lambda-sdk/test/lib/build-dummy-dev-mode-extension.js
@@ -3,7 +3,7 @@
 const path = require('path');
 const fsp = require('fs').promises;
 const mkdir = require('fs2/mkdir');
-const resolveDirZipBuffer = require('../utils/resolve-dir-zip-buffer');
+const resolveDirZipBuffer = require('../../../../test/utils/resolve-dir-zip-buffer');
 
 const zipFilename = path.resolve(__dirname, '../../dist/dummy-dev-mode.zip');
 const extensionDirname = path.resolve(__dirname, '../fixtures/dev-mode-extension');
