+++ node/test/utils/resolve-dir-zip-buffer.js
